task_service/requirements.txt
flask==2.3.1
openai==0.27.4
python-dotenv==0.21.0
tiktoken==0.3.3